PROSTHETIC-DESIGNER/backend/requirements.txt
Flask
Flask-Cors
Pillow
opencv-python-headless
numpy
